airflow/requirements.txt
apache-airflow-providers-google
pandas
beautifulsoup4
requests